game-image/pyproject.toml
[tool.poetry]
name = "game-ai-testing-platform-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
stockfish = "^3.28.0"
invoke = "^2.2.0"
python-dotenv = "^1.0.0"
python-chess = "^1.999"
numpy = "^1.26.4"
authlib = "^1.3.0"
gitpython = "^3.1.42"
gevent = "^24.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
